requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.1.2
certifi==2025.8.3
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
jinja2==3.1.6
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
pytz==2024.1
itsdangerous==2.1.2
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
